completely remove was-wrap library
